Refactor accounts namespace
