+++ nautilus_core/infrastructure/src/sql/cache_database.rs
@@ -22,7 +22,7 @@ use log::error;
 use nautilus_common::cache::database::CacheDatabaseAdapter;
 use nautilus_core::nanos::UnixNanos;
 use nautilus_model::{
-    account::any::AccountAny,
+    accounts::any::AccountAny,
     identifiers::{
         AccountId, ClientId, ClientOrderId, ComponentId, InstrumentId, PositionId, StrategyId,
         VenueOrderId,
