+++ nautilus_core/model/src/accounts/cash.rs
@@ -22,7 +22,7 @@ use std::{
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    account::base::{Account, BaseAccount},
+    accounts::base::{Account, BaseAccount},
     enums::{AccountType, LiquiditySide, OrderSide},
     events::{account::state::AccountState, order::filled::OrderFilled},
     identifiers::AccountId,
@@ -227,7 +227,7 @@ mod tests {
     use rstest::rstest;
 
     use crate::{
-        account::{base::Account, cash::CashAccount, stubs::*},
+        accounts::{base::Account, cash::CashAccount, stubs::*},
         enums::{AccountType, LiquiditySide, OrderSide},
         events::account::{state::AccountState, stubs::*},
         identifiers::{position_id::PositionId, AccountId},
