+++ nautilus_core/model/src/lib.rs
@@ -28,7 +28,7 @@
 //! - `python`: Enables Python bindings from `pyo3`
 //! - `stubs`: Enables type stubs for use in testing scenarios
 
-pub mod account;
+pub mod accounts;
 pub mod currencies;
 pub mod data;
 pub mod enums;
